refactor: Move version-command
